# [HUDI-8631] Support of `hoodie.populate.meta.fields` for Flink append mode