@geserdugarov 2024-12-20
I will restart Azure CI, but we have a problem with timeout:
https://dev.azure.com/apachehudi/hudi-oss-ci/_build/results?buildId=2536&view=results
All jobs are successful, but CI failed. I suppose due to running time of "UT spark-datasource Java Tests & DDL", 1 h 29 min 56 s, which is close to 1 h 30 min timeout. Could we increase timeout a little bit?